+++ changelog.md
@@ -30,3 +30,4 @@
 - blog stuff
 - yarn fix
 - Final fixes
+- cool mouse thing